GDA-5773 I11: Replace explicit "module load java" with call to dls-config/bin/loadjava.sh
Most of the modified scripts seem to be designed to run on workstations, so
they don't need to setup Environment Modules.

The exceptions to this are GDA_{Start,Stop}Servers, which run on i11-control.

startgda doesn't need to load Java at all; all the scripts it invokes do that.